requirements.txt
asgiref==3.8.1
Django==4.2
sqlparse==0.5.1
psycopg2-binary==2.9.6
djangorestframework==3.14.0
django-environ==0.11.2
gunicorn==22.0.0
whitenoise==6.7.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
